Add files via upload
The program asks the user to make a password. If the user's password is under 10 characters 5 points are reduced for every character. If the user's password doesn't contain a capital letter, a lowercase letter, and a number then 20 points is deducted. If the user has more than 2 consecutive letters or numbers that are the same then 20 points will also be deducted. If the user's points are 30 or more the user will be asked to try again.